% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFetchScript.R
\name{AnAge}
\alias{AnAge}
\title{Returns ageing and life history information in animals, including extensive longevity records, from HAGR AnAge}
\usage{
AnAge()
}
\value{
age, biblio, cysteine, development, habitat, links, metabolism, names, refs
}
\description{
Returns ageing and life history information in animals, including extensive longevity records, from HAGR AnAge
}
