% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anAge_age.R
\docType{data}
\name{anAge_age}
\alias{anAge_age}
\title{AnAge Phenotypes}
\format{
A table with 4766 rows and 9 variables:
\describe{
  \item{hagrid.name_common}{Common name for animal}
  \item{imr}{Infant Mortality Rate } 
  \item{mrdt}{Gene Entrez ID}
  \item{tmax}{Maximum Longevity}
  \item{size_sample}{Sample Size}
  \item{quality}{Data Quality}
  \item{origin}{Captivity or Wild}
  \item{biblioid}{ID for further information on data source, refering to anAge_biblio table}
}
}
\source{
\url{https://genomics.senescence.info/species/}
}
\usage{
anAge_age
}
\description{
AnAge table with species common names and associated phenotypes
}
\keyword{datasets}
