% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anAge_biblio.R
\docType{data}
\name{anAge_biblio}
\alias{anAge_biblio}
\title{AnAge Bibliographies}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1311 rows and 13 columns.
}
\source{
\url{https://genomics.senescence.info/species/}
}
\usage{
anAge_biblio
}
\description{
AnAge table with reference information for all entries
}
\keyword{datasets}
