% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anAge_development.R
\docType{data}
\name{anAge_development}
\alias{anAge_development}
\title{AnAge Development}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4741 rows and 13 columns.
}
\source{
\url{https://genomics.senescence.info/species/}
}
\usage{
anAge_development
}
\description{
AnAge table with developmental information for species
}
\keyword{datasets}
