% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anAge_metabolism.R
\docType{data}
\name{anAge_metabolism}
\alias{anAge_metabolism}
\title{AnAge Metabolism}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 707 rows and 4 columns.
}
\source{
\url{https://genomics.senescence.info/species/}
}
\usage{
anAge_metabolism
}
\description{
AnAge table with metabolism information for species
}
\keyword{datasets}
