% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anAge_names.R
\docType{data}
\name{anAge_names}
\alias{anAge_names}
\title{AnAge Names}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4766 rows and 15 columns.
}
\source{
\url{https://genomics.senescence.info/species/}
}
\usage{
anAge_names
}
\description{
AnAge table with nomenclature and naming information for species
}
\keyword{datasets}
