% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellAge_signatures.R
\docType{data}
\name{cellAge_signatures}
\alias{cellAge_signatures}
\title{CellAge Gene Signatures}
\format{
A table with 1259 rows and 7 variables:
\describe{
  \item{gene_symbol}{Gene symbol}
  \item{gene_name}{Gene name } 
  \item{entrez_id}{Gene Entrez ID}
  \item{total}{Total number of genes in study}
  \item{overexp}{Number of genes overexpressed}
  \item{underexp}{Number of genes underexpressed}
  \item{p_value}{p-value}
}
}
\source{
\url{https://genomics.senescence.info/cells/signatures.php}
}
\usage{
cellAge_signatures
}
\description{
Database of human genes that are either over-expressed or under-expressed during replicative senescence of human cells
}
\keyword{datasets}
