% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFetchScript.R
\name{DrugAge}
\alias{DrugAge}
\title{Returns all the tables in the DrugAge database}
\usage{
DrugAge()
}
\value{
biblio, compound_interactions, compound_synonyms, compounds, drugs
}
\description{
Returns all the tables in the DrugAge database
}
