% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugAge.R
\name{drugsCompare}
\alias{drugsCompare}
\title{Compare an input list of compounds with drugAge compounds}
\usage{
drugsCompare(drugList)
}
\arguments{
\item{druglist}{Character vector containing compound names}
}
\value{
The percentage match and all the matching drugs
}
\description{
Compare an input list of compounds with drugAge compounds
}
\examples{
myCompounds <- c('metformins','aspirin','blueberry','ethanol');
drugsCompare(myCompounds)
75 \% match to your list!
[1,] "ethanol"
[2,] "aspirin"
[3,] "metformin"
}
