% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFetchScript.R
\name{GenAgeHuman}
\alias{GenAgeHuman}
\title{Returns all data assocated with ageing in humans from HAGR GenAge human}
\usage{
GenAgeHuman()
}
\value{
Hbiblio, HFeatures, Hgene_biblio, hGO, Hhomolog, Hinteractions, Hlinks, Hmicroarray, Hnames, Horthologs, Hrefs and Hsequences
}
\description{
Returns all data assocated with ageing in humans from HAGR GenAge human
}
