% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genAgeHuman_features.R
\docType{data}
\name{genAgeHuman_features}
\alias{genAgeHuman_features}
\title{GenAge Human Genes}
\format{
A table with 307 rows and 8 variables:
\describe{
  \item{id}{Primary id of gene}
  \item{hagrid.gene_symbol}{Gene symbol linked to HAGR ID} 
  \item{why}{Reasoning behind including gene in GenAge}
  \item{band}{Cytogenetic band}
  \item{location_start}{Cytogenetic start location}
  \item{location_end}{Cytogenetic end location}
  \item{orientation}{Plus or Minus strand}
  \item{phenotype}{Potential phenotypic relvance to ageing}
}
}
\source{
\url{https://genomics.senescence.info/genes/query.php}
}
\usage{
genAgeHuman_features
}
\description{
Database of genes related to ageing in Humans
}
\keyword{datasets}
