% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genAgeHuman_gene_bio.R
\docType{data}
\name{genAgeHuman_gene_bio}
\alias{genAgeHuman_gene_bio}
\title{GenAge Gene References}
\format{
A table with 4460 rows and 4 variables:
\describe{
  \item{hagr_id}{Primary hagr id of gene}
  \item{year}{Year of primary publication} 
  \item{pubmed}{Pubmed ID of publication}
  \item{gene_symbol}{Gene Symbol}
}
}
\source{
\url{https://genomics.senescence.info/genes/query.php}
}
\usage{
genAgeHuman_gene_bio
}
\description{
Publication information for GenAge Human Genes
}
\keyword{datasets}
