% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genAgeHuman_names.R
\docType{data}
\name{genAgeHuman_names}
\alias{genAgeHuman_names}
\title{GenAge Gene Information}
\format{
A table with 307 rows and 5 variables:
\describe{
  \item{id_hagr}{Primary hagr id of gene}
  \item{alias}{Alternative names for gene} 
  \item{name_common}{Common name for gene}
  \item{entrez_id}{entrez_id}
}
}
\source{
\url{https://genomics.senescence.info/genes/query.php}
}
\usage{
genAgeHuman_names
}
\description{
Further information for genes in the GenAge Human Database
}
\keyword{datasets}
