% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFetchScript.R
\name{GenAgeModels}
\alias{GenAgeModels}
\title{Returns all data assocated with ageing in models from HAGR GenAge models}
\usage{
GenAgeModels()
}
\value{
Mbiblio, Morthologs, Mlongevity, Mgene_biblio, Mmodels
}
\description{
Returns all data assocated with ageing in models from HAGR GenAge models
}
