% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAGRR-package.R
\docType{package}
\name{HAGRR-package}
\alias{HAGRR}
\alias{HAGRR-package}
\title{HAGRR: Human Ageing Resources (HAGR)}
\description{
R data package for the Human Ageing Resources Databases
    https://genomics.senescence.info/
}
\keyword{internal}
